requirements.txt
Django==4.1.4
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
requests==2.28.1
